Reorganise files and add Modal
