+++ src/components/Button/Button.js
@@ -1,0 +1,11 @@
+import styles from './button.module.css'
+
+function Button({ children, onClick }) {
+  return (
+    <button className={styles.button} onClick={onClick}>
+      {children}
+    </button>
+  )
+}
+
+export default Button
