+++ src/components/Button/button.module.css
@@ -1,0 +1,14 @@
+.button {
+  background: #e80a4f 0% 0% no-repeat padding-box;
+  border-radius: 35px;
+  border: none;
+  padding: 20px 60px;
+  color: white;
+  font-size: 26px;
+  font-weight: bold;
+  margin-top: 60px;
+}
+.button:hover {
+  cursor: pointer;
+  background: #c00641;
+}
