+++ src/components/Modal/Modal.js
@@ -1,0 +1,11 @@
+import styles from './modal.module.css'
+
+const Modal = ({ children, isOpen }) => {
+  return (
+    <div className={`${styles.wrapper} ${isOpen ? '' : styles.hide}`}>
+      {children}
+    </div>
+  )
+}
+
+export default Modal
