+++ src/components/Modal/modal.module.css
@@ -1,0 +1,19 @@
+.wrapper {
+  position: absolute;
+  left: 0;
+  right: 0;
+  top: 0;
+  bottom: 0;
+  margin: 150px auto auto auto;
+
+  border-radius: 20px;
+  box-shadow: 0px 12px 30px #00000099;
+  background-color: #fff;
+  height: fit-content;
+  width: fit-content;
+  padding: 10px;
+}
+
+.hide {
+  display: none;
+}
