+++ src/components/domain/RestartModal/restartModal.module.css
@@ -1,0 +1,20 @@
+.wrapper {
+  padding: 30px;
+}
+
+.info {
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+}
+
+.text {
+  font-size: 32px;
+  font-weight: bold;
+  margin: 0;
+}
+
+.buttonWrapper {
+  display: flex;
+  justify-content: center;
+}
