+++ src/components/domain/Welcome/Welcome.js
@@ -1,5 +1,6 @@
-import { ReactComponent as Logo } from '../../assets/icons/logo.svg'
+import { ReactComponent as Logo } from '../../../assets/icons/logo.svg'
 import styles from './welcome.module.css'
+import Button from '../../Button/Button'
 
 function Welcome({ startGame }) {
   const handleStartGame = () => startGame('game')
@@ -8,9 +9,7 @@ function Welcome({ startGame }) {
     <div className={styles.content}>
       <Logo className={styles.logo} />
       <h1 className={styles.gameName}>MeMemory</h1>
-      <button className={styles.startButton} onClick={handleStartGame}>
-        Comenzar
-      </button>
+      <Button onClick={handleStartGame}>Comenzar</Button>
     </div>
   )
 }
